+++ config/jest/default.js
@@ -1,9 +1,17 @@
+const path = require('path');
+
+const globalLibDir = path.resolve(process.execPath, '../../lib/node_modules');
+
 module.exports = {
   collectCoverage: true,
   collectCoverageFrom: [
     'src/**/*.{js,jsx}',
   ],
   coverageDirectory: '__coverage__',
+  moduleDirectories: [
+    'node_modules',
+    globalLibDir,
+  ],
   moduleNameMapper: {
     '\\.(scss|css)$': 'identity-obj-proxy',
   },