+++ config/webpack/app-base.js
@@ -13,6 +13,7 @@ const moment = require('moment');
 const path = require('path');
 const { StatsWriterPlugin } = require('webpack-stats-plugin');
 const webpack = require('webpack');
+const WorkboxPlugin = require('workbox-webpack-plugin');
 
 /**
  * Creates a new Webpack config object, and performs some auxiliary operations
@@ -39,6 +40,10 @@ const webpack = require('webpack');
  *  include some polyfills we consider obligatory. If a string or an array is
  *  passed in, it is assigned to the "main" entry point, and the "polyfills"
  *  entry point will be added to it.
+ *
+ * @param {Boolean|Object} ops.workbox Adds InjectManifest plugin from Workbox,
+ *  with given options, if the argument is Object, or default ones, if it is any
+ *  other truly value.
  *
  * @param {Boolean} ops.keepBuildInfo Optional. If `true` and a build info file
  *  from a previous build is found, the factory will use that rather than
@@ -73,6 +78,9 @@ module.exports = function configFactory(ops) {
 
       /* Build timestamp. */
       timestamp: now.utc().toISOString(),
+
+      /* `true` if client-side code should setup a service worker. */
+      useServiceWorker: Boolean(o.workbox),
     };
     fs.writeFileSync(buildInfoUrl, JSON.stringify(buildInfo));
   }
@@ -86,9 +94,33 @@ module.exports = function configFactory(ops) {
   }
 
   entry.polyfills = _.union(entry.polyfills, [
-    'babel-polyfill',
+    '@babel/polyfill',
     'nodelist-foreach-polyfill',
   ]);
+
+  const plugins = [
+    new MiniCssExtractPlugin({
+      chunkFilename: `[name]-${now.valueOf()}.css`,
+      filename: `[name]-${now.valueOf()}.css`,
+    }),
+    new webpack.DefinePlugin({
+      BUILD_INFO: JSON.stringify(buildInfo),
+    }),
+    new StatsWriterPlugin({
+      filename: '__stats__.json',
+    }),
+  ];
+
+  /* Adds InjectManifest plugin from WorkBox, if opted to. */
+  if (o.workbox) {
+    if (!_.isObject(o.workbox)) o.workbox = {};
+    plugins.push(new WorkboxPlugin.InjectManifest({
+      importWorkboxFrom: 'local',
+      swSrc: path.resolve(__dirname, '../workbox/default.js'),
+      ...o.workbox,
+      swDest: '__service-worker.js',
+    }));
+  }
 
   return {
     context: o.context,
@@ -104,18 +136,7 @@ module.exports = function configFactory(ops) {
       path: path.resolve(__dirname, o.context, 'build'),
       publicPath: `${o.publicPath}/`,
     },
-    plugins: [
-      new MiniCssExtractPlugin({
-        chunkFilename: `[name]-${now.valueOf()}.css`,
-        filename: `[name]-${now.valueOf()}.css`,
-      }),
-      new webpack.DefinePlugin({
-        BUILD_INFO: JSON.stringify(buildInfo),
-      }),
-      new StatsWriterPlugin({
-        filename: '__stats__.json',
-      }),
-    ],
+    plugins,
     resolve: {
       alias: {
         /* Aliases to JS an JSX files are handled by Babel. */
@@ -137,7 +158,7 @@ module.exports = function configFactory(ops) {
         ],
         loader: 'file-loader',
         options: {
-          outputPath: '/fonts/',
+          outputPath: 'fonts/',
           publicPath: `${o.publicPath}/fonts`,
         },
       }, {
@@ -147,7 +168,8 @@ module.exports = function configFactory(ops) {
         loader: 'babel-loader',
         options: {
           babelrc: false,
-          forceEnv: o.babelEnv,
+          configFile: false,
+          envName: o.babelEnv,
           presets: ['topcoder-react-utils/config/babel/webpack'],
         },
       }, {
@@ -155,7 +177,7 @@ module.exports = function configFactory(ops) {
         test: /\.(gif|jpe?g|png)$/,
         loader: 'file-loader',
         options: {
-          outputPath: '/images/',
+          outputPath: 'images/',
           publicPath: `${o.publicPath}/images`,
         },
       }, {