+++ src/shared/utils/index.js
@@ -1,10 +1,12 @@
+import path from 'path';
+
 import config from './config';
 import * as isomorphy from './isomorphy';
 import * as redux from './redux';
 import * as webpack from './webpack';
 
-let juUrl = isomorphy.isProdBuild() ? 'prod' : 'dev';
-juUrl = `topcoder-react-utils/dist/${juUrl}/shared/utils/jest`;
+const juUrl = module.webpackPolyfill ? './shared/utils/jest'
+  : path.resolve(__dirname, './jest');
 
 const JU = isomorphy.isServerSide() ? webpack.requireWeak(juUrl) : null;